pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arabic-rag-chatbot"
version = "1.0.0"
description = "Production-ready Arabic RAG chatbot with conversation memory using 2025 benchmarks"
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["arabic", "nlp", "rag", "chatbot", "langchain", "qdrant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-huggingface>=0.0.1",
    "langchain-openai>=0.0.5",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.3.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "camel-tools>=1.5.0",
    "pyarabic>=0.6.15",
    "fastapi>=0.109.0",
    "python-multipart>=0.0.6",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "html5lib>=1.1",
    "markdown-it-py>=3.0.0",
    "python-docx>=1.1.0",
    "chardet>=5.2.0",
    "python-magic>=0.4.0",
    "Pillow>=10.0.0",
    "pypdf>=4.0.0",
    "requests>=2.31.0",
    "lxml>=5.0.0",
    "httpx>=0.27.0",
    "firecrawl-py>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
evaluation = [
    "ragas>=0.1.0",
    "datasets>=2.16.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "performance: Performance benchmarks",
]
addopts = "--strict-markers --cov=src --cov-report=term-missing --cov-report=html --cov-fail-under=80 -v"
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py", "**/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
